Split up scripting namespace
