+++ src/script/api/buffer.rs
@@ -1,0 +1,37 @@
+use std::fmt;
+
+use crate::{
+    editing::Id,
+    input::{commands::CommandHandlerContext, KeymapContext},
+};
+
+use super::Api;
+
+#[apigen::ns]
+pub struct BufferApiObject {
+    api: Api,
+    pub id: Id,
+}
+
+impl fmt::Debug for BufferApiObject {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "<Buffer #{}>", self.id)
+    }
+}
+
+#[apigen::ns_impl]
+impl BufferApiObject {
+    pub fn new(api: Api, id: Id) -> Self {
+        Self { api, id }
+    }
+
+    #[property]
+    #[rpc(passing(self.id))]
+    pub fn name(context: &mut CommandHandlerContext, id: Id) -> Option<String> {
+        if let Some(buf) = context.state().buffers.by_id(id) {
+            Some(format!("{:?}", buf.source()))
+        } else {
+            None
+        }
+    }
+}
