+++ src/script/api/core.rs
@@ -1,7 +1,6 @@
 use std::{fmt, io};
 
 use crate::{
-    editing::Id,
     input::{
         commands::CommandHandlerContext, keys::KeysParsable, maps::UserKeyHandler, KeymapContext,
         RemapMode,
@@ -9,7 +8,7 @@ use crate::{
     script::fns::ScriptingFnRef,
 };
 
-use super::{Api, Fns};
+use super::{current::CurrentObjects, Api, Fns};
 
 #[apigen::ns]
 #[derive(Clone)]
@@ -74,61 +73,4 @@ fn create_user_keyhandler(f: ScriptingFnRef) -> Box<UserKeyHandler> {
             })
             .join_interruptably(&mut ctx)
     })
-}
-
-#[apigen::ns]
-pub struct CurrentObjects {
-    api: Api,
-}
-
-#[apigen::ns_impl]
-impl CurrentObjects {
-    pub fn new(api: Api) -> Self {
-        Self { api }
-    }
-
-    #[rpc]
-    fn buffer_id(context: &mut CommandHandlerContext) -> Id {
-        context.state().current_buffer().id()
-    }
-
-    #[property]
-    pub fn buffer(&self) -> BufferApiObject {
-        BufferApiObject::new(self.api.clone(), self.buffer_id())
-    }
-}
-
-impl fmt::Debug for CurrentObjects {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "<CurrentObjects>")
-    }
-}
-
-#[apigen::ns]
-pub struct BufferApiObject {
-    api: Api,
-    pub id: Id,
-}
-
-impl fmt::Debug for BufferApiObject {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "<Buffer #{}>", self.id)
-    }
-}
-
-#[apigen::ns_impl]
-impl BufferApiObject {
-    pub fn new(api: Api, id: Id) -> Self {
-        Self { api, id }
-    }
-
-    #[property]
-    #[rpc(passing(self.id))]
-    pub fn name(context: &mut CommandHandlerContext, id: Id) -> Option<String> {
-        if let Some(buf) = context.state().buffers.by_id(id) {
-            Some(format!("{:?}", buf.source()))
-        } else {
-            None
-        }
-    }
 }
